Owner commands will return error messages for users that can't use those commands.

+++ src/commands/Owner Only/say.js
@@ -32,6 +32,9 @@ class SayMessage extends Command {
     }
 
     async exec(message, { thing }) {
+        
+        if (message.author.id !== "594853883742912512") return message.reply("**Only `Bot Owner` can run this command.**");
+
         message.channel.send(thing);
         message.delete();
     }
